requirements.txt
openai-agents>=0.1.0
python-dotenv>=1.0.0
rich>=13.0.0